requirements.txt

# Phase 10 — Nano Banana texture factory
google-genai>=0.3.0
python-dotenv>=1.0.0
